LayoutTests/http/tests/loading/preload-no-store-frame-src.html

<html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpResourceResponseMIMETypes();
</script>
<script src="../../resources/js-test-pre.js"></script>
<script src="http://127.0.0.1:8000/resources/slow-script.pl?delay=100"></script>
<script>
    shouldBeTrue('internals.isPreloaded("/navigation/resources/no-store-frame.php");');
</script>
</head>
<body>
<p>Tests that 'no-store' iframes are preloaded.</p>
<iframe src="/navigation/resources/no-store-frame.php"></iframe>
</body>
</html>
