LayoutTests/http/tests/loading/preload-no-store-frame-src-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
js-test-pre.js has MIME type application/x-javascript
slow-script.pl has MIME type application/javascript
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
no-store-frame.php has MIME type text/html
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
PASS internals.isPreloaded("/navigation/resources/no-store-frame.php"); is true
Tests that 'no-store' iframes are preloaded.
